Move tenant fetching logic from handler to tenant_fetcher
